Updates
Converted all output in CreateProject.ps1, including errors, into
message boxes.  This allows the script to be run without having to leave
the PowerShell console open to read any error messages.

Added code to create a "New Project Here" shortcut in each subject
folder.  This is a dynamically generated shortcut that is different in
each folder, and will be overwritten every time the
VerifyFolderStructure.ps1 script runs.  It launches the
CreateProject.ps1 script with the proper path parameter automatically;
no need for an in-between batch file.

Added log file output, via a Logging module that I previously wrote, to
keep the changes to the VerifyFolderStructure.ps1 script to a minimum.

Added some test code for calling msg.exe as an output option.

Separated the HTML generation code into its own function, making it
easier to modify later, and making the code that checks the folder tree
easier to read.